feat: implement responsive design
The previous implementation did not work
well with smaller screen sizes. There were
issues with the width of the placeholder
text in the input. There were also issues
with the width of the preview window for the
prefix tree that would cut off certain text.

In addition to fixing responsive design issues,
a button was added to make it more clear that
the input needs to be submitted for the word
to show up in the tree.